memory finder separate file
